+++ servant-server/src/Servant/Server/Internal.hs
@@ -52,7 +52,7 @@ import           Servant.API                 ((:<|>) (..), (:>), Capture,
                                               Verb, ReflectMethod(reflectMethod),
                                               IsSecure(..), Header,
                                               QueryFlag, QueryParam, QueryParams,
-                                              Raw, RemoteHost, ReqBody, Vault,
+                                              Raw(..), RemoteHost, ReqBody, Vault,
                                               WithNamedConfig)
 import           Servant.API.ContentTypes    (AcceptHeader (..),
                                               AllCTRender (..),
@@ -354,18 +354,23 @@ instance (KnownSymbol sym, HasServer sublayout config)
 --
 -- Example:
 --
--- > type MyApi = "images" :> Raw
+-- > type MyApi = "images" :> Raw IO Application
 -- >
 -- > server :: Server MyApi
 -- > server = serveDirectory "/var/www/images"
-instance HasServer Raw config where
+class ToRawApplication a where
+  toRawApplication :: a -> Application
 
-  type ServerT Raw m = Application
+instance ToRawApplication Application where
+  toRawApplication = id
+
+instance ToRawApplication a => HasServer (Raw m a) config where
+  type ServerT (Raw m a) n = Raw n a
 
   route Proxy _ rawApplication = LeafRouter $ \ request respond -> do
     r <- runDelayed rawApplication
     case r of
-      Route app   -> app request (respond . Route)
+      Route (Raw app)   -> (toRawApplication app) request (respond . Route)
       Fail a      -> respond $ Fail a
       FailFatal e -> respond $ FailFatal e