+++ servant-client/CHANGELOG.md
@@ -2,6 +2,7 @@ HEAD
 ----
 
 * Use the `text` package instead of `String`.
+* Update `instance HasClient Raw` for servant Raw changes
 * Support for the `HttpVersion`, `IsSecure`, `RemoteHost` and `Vault` combinators
 * Added support for `path` on `BaseUrl`.
 * `client` now takes an explicit `Manager` argument.